+++ src/theme/mod.rs
@@ -7,8 +7,10 @@ use crate::output::render;
 mod ui_styles;
 pub use self::ui_styles::{DateAge, UiStyles};
 
+pub mod key_registry;
+
 mod lsc;
-pub use self::lsc::LSColors;
+pub use self::lsc::{LSColors, render_style_to_lx};
 
 mod default_theme;